Tests with ugly switch/case nearly completed

+++ src/Visutron900PlusInterface.Adapter/DTOs/RefraktionDataIn.cs
@@ -5,14 +5,14 @@
         public double SphäreFernRechts { get; set; }
         public double SphäreNahRechts { get; set; }
         public double ZylinderRechts { get; set; }
-        public double AchseRechts { get; set; }
+        public int AchseRechts { get; set; }
         public double PrismaRechts { get; set; }
         public double PupillendistanzRechts { get; set; }
 
         public double SphäreFernLinks { get; set; }
         public double SphäreNahLinks { get; set; }
         public double ZylinderLinks { get; set; }
-        public double AchseLinks { get; set; }
+        public int AchseLinks { get; set; }
         public double PrismaLinks { get; set; }
         public double PupillendistanzLinks { get; set; }
